refactor: 메인페이지 스트리밍 렌더링 도입
- 메인 페이지 모임 목록 섹션에 Suspense 및 스트리밍 적용
- 데이터 페칭 블로킹 제거를 통한 초기 로딩 성능 개선
