+++ src/app/main/components/MyGatheringSection/index.tsx
@@ -1,7 +1,9 @@
 'use client';
 
-import { useAuth } from '@/hooks/useAuth';
+import { GatheringSectionSkeleton } from '@/components/GatheringSectionSkeleton';
 import { SuspenseBoundary } from '@/components/SuspenseBoundary';
+import { useAuth } from '@/hooks/useAuth';
+
 import { MyGatheringList } from './MyGatheringList';
 
 export function MyGatheringSection() {
@@ -13,7 +15,8 @@ export function MyGatheringSection() {
 
   return (
     <SuspenseBoundary
-      pendingFallback={<div className='p-4 text-center text-gray-500'>모임을 불러오는 중입니다...</div>}
+      // 🟢 단순 텍스트 대신 새로 만든 고퀄리티 스켈레톤을 적용합니다.
+      pendingFallback={<GatheringSectionSkeleton />}
       errorFallback={(error, reset) => (
         <div className='p-4 text-center text-red-500'>
           <p>모임을 불러올 수 없습니다.</p>
